+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added chart showing wins and losses per map to the Trends page
 - A warning is now shown when trying to log a match in a past competitive season [#17](https://github.com/cheshire137/competiwatch-desktop/issues/17)
 - App title now updates to reflect the selected account, season, and page [#19](https://github.com/cheshire137/competiwatch-desktop/issues/19)
+- Fixed bug where new competitive seasons couldn't be added from the Manage Seasons page
 
 ## Version 0.1.2